+++ apps/markup-ai/src/locations/Dialog/dialogTypes.ts
@@ -1,0 +1,11 @@
+import { StyleScores } from '@markupai/toolkit';
+export interface RewriteDialogParams {
+  startRewrite: boolean;
+  fieldId: string;
+  original: string;
+  originalScore?: number | null;
+}
+
+export interface MoreDetailsDialogParams {
+  scores: StyleScores;
+}